+++ packages/providers/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./fake.js";
 export * from "./defaults.js";
 export * from "./env.js";
+export * from "./profileRegistry.js";
 export * from "./openaiCompatible.js";
 export * from "./pricing.js";
 export * from "./registry.js";